[#15] feat : 로컬에서 스웨거 문서 테스트 성공
