+++ src/main/java/com/titanic/fork/config/WebConfig.java
@@ -34,7 +34,7 @@ public class WebConfig implements WebMvcConfigurer {
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
 
-        String[] excludePathPatterns = new String[]{"/account/**"};
+        String[] excludePathPatterns = new String[]{"/account/**", "/**"};
 
         registry.addInterceptor(loginInterceptor())
                 .addPathPatterns(ALL_PATH)
