+++ src/main/java/com/titanic/fork/web/controller/RegisterController.java
@@ -2,6 +2,7 @@ package com.titanic.fork.web.controller;
 
 import com.titanic.fork.service.account.RegisterService;
 import com.titanic.fork.web.dto.request.account.RegisterWantDto;
+import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,6 +17,9 @@ public class RegisterController {
 
     private final RegisterService registerService;
 
+    @ApiOperation(value = "회원가입API",
+                    notes = "회원가입 성공 시 HttpStatus = 201(Created), Header에 Authorization에 JWT 토큰 응답 \n" +
+                            "회원가입 실패 시 HttpStatus = 500(Internal Sever Error)")
     @PostMapping("account")
     public ResponseEntity<Void> register(@RequestBody RegisterWantDto registerWantDto,
                                          HttpServletResponse response) {
